chore(cs): move duplicate exception to a single place
Refs: XRDDEV-2291
