+++ src/central-server/admin-service/infra-api-rest/src/main/java/org/niis/xroad/cs/admin/rest/api/openapi/InitializationApiController.java
@@ -36,15 +36,15 @@ import org.niis.xroad.cs.admin.api.exception.WeakPinException;
 import org.niis.xroad.cs.admin.api.service.InitializationService;
 import org.niis.xroad.cs.admin.rest.api.converter.model.InitialServerConfDtoConverter;
 import org.niis.xroad.cs.admin.rest.api.converter.model.InitializationStatusDtoConverter;
-import org.niis.xroad.cs.admin.rest.api.exception.ConflictException;
-import org.niis.xroad.cs.admin.rest.api.exception.InternalServerErrorException;
 import org.niis.xroad.cs.openapi.InitializationApi;
 import org.niis.xroad.cs.openapi.model.InitialServerConfDto;
 import org.niis.xroad.cs.openapi.model.InitializationStatusDto;
 import org.niis.xroad.restapi.config.audit.AuditEventMethod;
 import org.niis.xroad.restapi.config.audit.RestApiAuditEvent;
 import org.niis.xroad.restapi.openapi.BadRequestException;
+import org.niis.xroad.restapi.openapi.ConflictException;
 import org.niis.xroad.restapi.openapi.ControllerUtil;
+import org.niis.xroad.restapi.openapi.InternalServerErrorException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
