+++ src/common-domain/src/main/java/org/niis/xroad/restapi/exceptions/ErrorDeviation.java
@@ -4,17 +4,17 @@
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
  * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- *
+ * <p>
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- *
+ * <p>
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- *
+ * <p>
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -31,6 +31,7 @@ import java.util.List;
  * An error that can't be ignored
  */
 public class ErrorDeviation extends Deviation {
+
     public ErrorDeviation(String code, List<String> metadata) {
         super(code, metadata);
     }
@@ -41,5 +42,37 @@ public class ErrorDeviation extends Deviation {
 
     public ErrorDeviation(String code) {
         super(code);
+    }
+
+    /**
+     * Creates new error deviation.
+     *
+     * @param code     error code
+     * @param metadata metadata
+     * @return new ErrorDeviation
+     */
+    public static ErrorDeviation newError(String code, List<String> metadata) {
+        return new ErrorDeviation(code, metadata);
+    }
+
+    /**
+     * Creates new error deviation.
+     *
+     * @param code         error code
+     * @param metadataItem metadata
+     * @return new ErrorDeviation
+     */
+    public static ErrorDeviation newError(String code, String metadataItem) {
+        return new ErrorDeviation(code, metadataItem);
+    }
+
+    /**
+     * Creates new error deviation.
+     *
+     * @param code error code
+     * @return new ErrorDeviation
+     */
+    public static ErrorDeviation newError(String code) {
+        return new ErrorDeviation(code);
     }
 }
