+++ src/common-domain/src/main/java/org/niis/xroad/restapi/openapi/BadRequestException.java
@@ -26,6 +26,7 @@
 package org.niis.xroad.restapi.openapi;
 
 import org.niis.xroad.restapi.exceptions.DeviationAwareException;
+import org.niis.xroad.restapi.exceptions.DeviationAwareRuntimeException;
 import org.niis.xroad.restapi.exceptions.ErrorDeviation;
 import org.niis.xroad.restapi.exceptions.WarningDeviation;
 import org.springframework.http.HttpStatus;
@@ -38,7 +39,7 @@ import java.util.Collection;
  * Results in http 400 BAD_REQUEST
  */
 @ResponseStatus(value = HttpStatus.BAD_REQUEST)
-public class BadRequestException extends OpenApiException {
+public class BadRequestException extends DeviationAwareRuntimeException {
 
     public BadRequestException(DeviationAwareException e) {
         super(e, e.getErrorDeviation(), e.getWarningDeviations());
@@ -57,14 +58,6 @@ public class BadRequestException extends OpenApiException {
 
     public BadRequestException(String msg, Throwable t, ErrorDeviation errorDeviation) {
         super(msg, t, errorDeviation);
-    }
-
-    public BadRequestException(String msg, Throwable t) {
-        super(msg, t);
-    }
-
-    public BadRequestException(Throwable t) {
-        super(t);
     }
 
     public BadRequestException(Throwable t, ErrorDeviation errorDeviation,
