+++ src/security-server/admin-service/application/src/test/java/org/niis/xroad/securityserver/restapi/openapi/BackupsApiControllerTest.java
@@ -37,6 +37,7 @@ import org.niis.xroad.restapi.common.backup.exception.BackupInvalidFileException
 import org.niis.xroad.restapi.exceptions.DeviationCodes;
 import org.niis.xroad.restapi.exceptions.WarningDeviation;
 import org.niis.xroad.restapi.openapi.BadRequestException;
+import org.niis.xroad.restapi.openapi.InternalServerErrorException;
 import org.niis.xroad.restapi.openapi.ResourceNotFoundException;
 import org.niis.xroad.restapi.service.UnhandledWarningsException;
 import org.niis.xroad.securityserver.restapi.openapi.model.Backup;
