+++ src/security-server/admin-service/application/src/test/java/org/niis/xroad/securityserver/restapi/openapi/InitializationApiControllerTest.java
@@ -28,6 +28,8 @@ package org.niis.xroad.securityserver.restapi.openapi;
 import org.junit.Test;
 import org.mockito.Mockito;
 import org.niis.xroad.restapi.openapi.BadRequestException;
+import org.niis.xroad.restapi.openapi.ConflictException;
+import org.niis.xroad.restapi.openapi.InternalServerErrorException;
 import org.niis.xroad.restapi.service.UnhandledWarningsException;
 import org.niis.xroad.securityserver.restapi.dto.InitializationStatusDto;
 import org.niis.xroad.securityserver.restapi.openapi.model.InitialServerConf;
